add DummyOFunction_P

+++ scripts/libobjectivefunction.py
@@ -23,7 +23,7 @@ class DummyOFunction_P(pycalimero.IObjectiveFunction):
     def __init__(self):
         pycalimero.IObjectiveFunction.__init__(self)
         self.setDataType("fac n", pycalimero.UINT, "0")
-        self.setDataType("iterations", pycalimero.DOUBLE, "1000")
+        self.setDataType("iterations", pycalimero.UINT, "1000")
         
     def eval(self,iterationparameters, observedparameters, objectivefunctionparameters):
         result = pycalimero.doublevector()
